Move identity to its own module
