+++ tower-layer/src/util/mod.rs
@@ -3,11 +3,10 @@
 use Layer;
 
 mod chain;
+mod identity;
 
 pub use self::chain::{Chain, ChainError};
-
-use std::error::Error;
-use std::fmt;
+pub use self::identity::Identity;
 
 /// An extension trait for `Layer`'s that provides a variety of convenient
 /// adapters.
@@ -26,46 +25,3 @@ pub trait LayerExt<S, Request>: Layer<S, Request> {
 }
 
 impl<T, S, Request> LayerExt<S, Request> for T where T: Layer<S, Request> {}
-
-/// A no-op middleware.
-///
-/// When wrapping a `Service`, the `Identity` layer returns the provided
-/// service without modifying it.
-#[derive(Debug, Default, Clone)]
-pub struct Identity {
-    _p: (),
-}
-
-impl Identity {
-    /// Create a new `Identity` value
-    pub fn new() -> Identity {
-        Identity { _p: () }
-    }
-}
-
-/// Decorates a `Service`, transforming either the request or the response.
-impl<S, Request> Layer<S, Request> for Identity
-where
-    S: Service<Request>,
-{
-    type Response = S::Response;
-    type Error = S::Error;
-    type LayerError = Never;
-    type Service = S;
-
-    fn layer(&self, inner: S) -> Result<Self::Service, Self::LayerError> {
-        Ok(inner)
-    }
-}
-
-#[derive(Debug)]
-/// An error that can never occur.
-pub enum Never {}
-
-impl fmt::Display for Never {
-    fn fmt(&self, _: &mut fmt::Formatter) -> fmt::Result {
-        match *self {}
-    }
-}
-
-impl ::std::error::Error for Never {}
